food-detector/requirements.txt
ultralytics>=8.3.0
coremltools>=8.0
opencv-python>=4.8.0
PyYAML>=6.0
mediapipe>=0.10.0
